+++ src/sophronia/app/time/page.js
@@ -2,9 +2,9 @@ import { getBackendURL } from 'lib/ApiFetcher';
 
 import { PageTitle } from 'components/ui/PageTitle';
 import { PageContent } from 'components/ui/PageContent';
-// import { ExportCSV } from 'components/ui/ExportCSV';
+import { ExportCSV } from 'components/ui/ExportCSV';
 import { StructuredTimeData } from 'components/StructuredData';
-import { Card } from 'components/Dashboard/Card';
+import { Card } from 'components/ui/Card';
 
 import { AllTimeChart } from './AllTimeChart';
 import { YearsTable } from './YearsTable';
@@ -68,7 +68,7 @@ export default async function Page() {
 				</Card>
 			</div>
 
-			{/* <ExportCSV data={data.results} filename={'alltime.csv'} /> */}
+			<ExportCSV data={data.results} filename={'alltime.csv'} />
 			<YearsTable data={data.results} id={'yearstable'} />
 		</>
 	);
