+++ src/sophronia/components/Dashboard/Datasource.jsx
@@ -1,5 +1,5 @@
 import { Button } from '../ui/Button';
-import { Card } from './Card';
+import { Card } from 'components/ui/Card';
 import { RiSeedlingLine } from 'react-icons/ri';
 
 /**
