+++ src/sophronia/components/Dashboard/LastUpdated.jsx
@@ -1,4 +1,4 @@
-import { Card } from './Card';
+import { Card } from 'components/ui/Card';
 
 /**
  * @file LastUpdated.jsx
